+++ woods/src/main/scala/com/github/gchudnov/woods/box/BasicBox.scala
@@ -3,4 +3,4 @@ package com.github.gchudnov.woods.box
 import com.github.gchudnov.swearwolf.util.Size
 import com.github.gchudnov.woods.{Box, BoxStyle}
 
-final case class BasicBox(size: Size, style: BoxStyle) extends Box
+private[woods] final case class BasicBox(size: Size, style: BoxStyle) extends Box
