--- woods/src/main/scala/com/github/gchudnov/woods/box/BoxException.scala
@@ -1,3 +1,0 @@
-package com.github.gchudnov.woods.box
-
-final class BoxException(message: String) extends RuntimeException(message)
