+++ woods/src/main/scala/com/github/gchudnov/woods/box/BoxSyntax.scala
@@ -1,7 +1,15 @@
 package com.github.gchudnov.woods.box
 
 import com.github.gchudnov.swearwolf.Screen
+import com.github.gchudnov.swearwolf.util.{Point, TextStyle}
+import com.github.gchudnov.woods.Box
+import com.github.gchudnov.woods.box.impl.BoxDrawer
 
-trait BoxSyntax {
+private[box] class BoxOps(private val screen: Screen) extends AnyVal {
+  def put(pt: Point, box: Box, textStyle: TextStyle = TextStyle.Empty): Either[Throwable, Unit] =
+    BoxDrawer.draw(screen)(pt, box, textStyle)
+}
+
+private[woods] trait BoxSyntax {
   implicit def boxOps(screen: Screen): BoxOps = new BoxOps(screen)
 }
