+++ woods/src/main/scala/com/github/gchudnov/woods/grid/BasicGrid.scala
@@ -3,4 +3,4 @@ package com.github.gchudnov.woods.grid
 import com.github.gchudnov.swearwolf.util.Size
 import com.github.gchudnov.woods.{Grid, GridStyle}
 
-final case class BasicGrid(size: Size, cell: Size, style: GridStyle) extends Grid
+private[woods] final case class BasicGrid(size: Size, cell: Size, style: GridStyle) extends Grid
