+++ woods/src/main/scala/com/github/gchudnov/woods/grid/GridSyntax.scala
@@ -1,7 +1,15 @@
 package com.github.gchudnov.woods.grid
 
 import com.github.gchudnov.swearwolf.Screen
+import com.github.gchudnov.swearwolf.util.{Point, TextStyle}
+import com.github.gchudnov.woods.Grid
+import com.github.gchudnov.woods.grid.impl.GridDrawer
 
-trait GridSyntax {
+private[grid] class GridOps(private val screen: Screen) extends AnyVal {
+  def put(pt: Point, grid: Grid, textStyle: TextStyle = TextStyle.Empty): Either[Throwable, Unit] =
+    GridDrawer.draw(screen)(pt, grid, textStyle)
+}
+
+private[woods] trait GridSyntax {
   implicit def gridOps(screen: Screen): GridOps = new GridOps(screen)
 }
