+++ woods/src/main/scala/com/github/gchudnov/woods/label/BasicLabel.scala
@@ -3,4 +3,4 @@ package com.github.gchudnov.woods.label
 import com.github.gchudnov.swearwolf.util.Size
 import com.github.gchudnov.woods.{AlignStyle, Label}
 
-final case class BasicLabel(size: Size, value: String, align: AlignStyle, isFill: Boolean) extends Label
+private[woods] final case class BasicLabel(size: Size, value: String, align: AlignStyle, isFill: Boolean) extends Label
