+++ woods/src/main/scala/com/github/gchudnov/woods/label/LabelSyntax.scala
@@ -1,7 +1,15 @@
 package com.github.gchudnov.woods.label
 
 import com.github.gchudnov.swearwolf.Screen
+import com.github.gchudnov.swearwolf.util.{Point, TextStyle}
+import com.github.gchudnov.woods.Label
+import com.github.gchudnov.woods.label.impl.LabelDrawer
 
-trait LabelSyntax {
+private[label] class LabelOps(private val screen: Screen) extends AnyVal {
+  def put(pt: Point, label: Label, textStyle: TextStyle = TextStyle.Empty): Either[Throwable, Unit] =
+    LabelDrawer.draw(screen)(pt, label, textStyle)
+}
+
+private[woods] trait LabelSyntax {
   implicit def labelOps(screen: Screen): LabelOps = new LabelOps(screen)
 }
