+++ woods/src/main/scala/com/github/gchudnov/woods/table/BasicTable.scala
@@ -2,4 +2,4 @@ package com.github.gchudnov.woods.table
 
 import com.github.gchudnov.woods.{Table, TableStyle}
 
-final case class BasicTable(data: Seq[Seq[Any]], style: TableStyle) extends Table
+private[woods] final case class BasicTable(data: Seq[Seq[Any]], style: TableStyle) extends Table
