--- woods/src/main/scala/com/github/gchudnov/woods/table/package.scala
@@ -1,14 +1,0 @@
-package com.github.gchudnov.woods
-
-import com.github.gchudnov.swearwolf.Screen
-import com.github.gchudnov.swearwolf.util.{ Point, TextStyle }
-import com.github.gchudnov.woods.table.impl.TableDrawer
-
-package object table {
-
-  class TableOps(private val screen: Screen) extends AnyVal {
-    def put(pt: Point, table: Table, textStyle: TextStyle = TextStyle.Empty): Either[Throwable, Unit] =
-      TableDrawer.draw(screen)(pt, table, textStyle)
-  }
-
-}
