+++ woods/src/main/scala/com/github/gchudnov/woods/text/CompiledRichText.scala
@@ -1,0 +1,5 @@
+package com.github.gchudnov.woods.text
+
+import com.github.gchudnov.woods.RichText
+
+private[woods] final case class CompiledRichText(bytes: Array[Byte]) extends RichText
