+++ woods/src/main/scala/com/github/gchudnov/woods/text/RichTextException.scala
@@ -1,3 +1,3 @@
 package com.github.gchudnov.woods.text
 
-final class RichTextException(message: String) extends RuntimeException(message)
+private[text] final class RichTextException(message: String) extends RuntimeException(message)
