+++ woods/src/main/scala/com/github/gchudnov/woods/text/RichTextFactory.scala
@@ -1,6 +1,6 @@
 package com.github.gchudnov.woods.text
 
-import com.github.gchudnov.woods.{CompiledRichText, RichText}
+import com.github.gchudnov.woods.{RichText}
 import com.github.gchudnov.woods.text.impl.{RichTextCompiler, RichTextParser, RichTextStyler}
 
 private[woods] object RichTextFactory {
