+++ woods/src/main/scala/com/github/gchudnov/woods/text/RichTextSyntax.scala
@@ -1,7 +1,14 @@
 package com.github.gchudnov.woods.text
 
 import com.github.gchudnov.swearwolf.Screen
+import com.github.gchudnov.swearwolf.util.Point
+import com.github.gchudnov.woods.RichText
 
-trait RichTextSyntax {
+private[text] class RichTextOps(private val screen: Screen) extends AnyVal {
+  def put(pt: Point, value: RichText): Either[Throwable, Unit] =
+    screen.put(pt, value.bytes)
+}
+
+private[woods] trait RichTextSyntax {
   implicit def richTextOps(screen: Screen): RichTextOps = new RichTextOps(screen)
 }
